+++ bindings/julia/src/boolean.jl
@@ -1,0 +1,2 @@
+_integer_to_boolean(value::Nothing) = nothing
+_integer_to_boolean(value::Integer)::Bool = Bool(value)